First stages of implementation
Started creating a basic codebase and methodology for using waylandpp.
Each wayland resource will have one implementation object alive for it
and will be valid for the duration of the resource.

Abstraction for this is made possible by inheriting from
`ObjectImplementationBase` and specifying the resource, and the
implementing class type as templated arguments.

Currently, the Wayland Shared Memory interfaces are mostly (not fully)
implemented. These include wl_shm, wl_shm_pool, and wl_buffer.