Cleaned up and improved `ObjectImplementationBase`
Learning more about waylandpp allowed me to clean up my base
implementation class. ObjectData is now redundant and removed. Instead,
object data is stored within the class, and the class is stored as a
pointer in the user_data of the underlying resource.

This also removes the very high possibility that std::bad_weak_ptr is
thrown due to using shared_from_this() during the creation of a
std::shared_ptr.