Implemented Event system and implemented wl_keyboard
Events in Moco are handled in a subscription based system. Something
wanting to get notified of an event, will "subscribe" to it using
`Events::Subscribe(Event)` and will get a subscription object back.
Upon destruction, the event will automatically be unsubscribed.

Things wanting to publish an event can call `Events::Publish(Event,
EventData)`, which will broadcast to all subscribed objects and pass
EventData to their callbacks.

Using this we can finish wl_keyboard, but we still need a so-called
backend to send keyboard events.